+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Add createBranch and branchName to test\_helpers.dart
+
 ## [2.5.15] - 2024-08-30
 
 ### Changed
@@ -108,6 +114,7 @@
 - `HeadMessage`: Allow to return head message also when not everything is committed.
 - `modified-files`
 
+[Unreleased]: https://github.com/inlavigo/gg_git/compare/2.5.15...HEAD
 [2.5.15]: https://github.com/inlavigo/gg_git/compare/2.5.14...2.5.15
 [2.5.14]: https://github.com/inlavigo/gg_git/compare/2.5.13...2.5.14
 [2.5.13]: https://github.com/inlavigo/gg_git/compare/2.5.12...2.5.13